+++ Luthor.Tests/LexerTests.cs
@@ -8,18 +8,11 @@ namespace Luthor.Tests
     [TestClass]
     public class LexerTests
     {
-        private Lexer lexer;
-
-        private void Setup(string source)
-        {
-            lexer = new Lexer(source);
-        }
-
         [TestMethod]
         public void NoSource_ReturnsEOF()
         {
             // Arrange.
-            Setup(string.Empty);
+            var lexer = new Lexer(string.Empty);
 
             // Act.
             var tokens = lexer.GetTokens();
@@ -33,7 +26,7 @@ namespace Luthor.Tests
         public void WithVaryingLineEndings_ReturnsCorrectLineCounts()
         {
             // Arrange.
-            Setup("  \r\r\n \n\r \n");
+            var lexer = new Lexer("  \r\r\n \n\r \n");
 
             // Act.
             var tokens = lexer.GetTokens(true);
@@ -43,10 +36,35 @@ namespace Luthor.Tests
         }
 
         [TestMethod]
+        public void WithOneOfEachKindOfToken_ReturnsCorrectTokens()
+        {
+            // Arrange.
+            var lexer = new Lexer("AB 01 +- 'AB' àë \n");
+
+            // Act.
+            var tokens = lexer.GetTokens(true);
+
+            // Assert.
+            var t = 0;
+            AssertToken(tokens, ref t, TokenTypes.Letters, "AB");
+            AssertToken(tokens, ref t, TokenTypes.Whitespace, " ");
+            AssertToken(tokens, ref t, TokenTypes.Digits, "01");
+            AssertToken(tokens, ref t, TokenTypes.Whitespace, " ");
+            AssertToken(tokens, ref t, TokenTypes.Symbols, "+-");
+            AssertToken(tokens, ref t, TokenTypes.Whitespace, " ");
+            AssertToken(tokens, ref t, TokenTypes.String, "'AB'");
+            AssertToken(tokens, ref t, TokenTypes.Whitespace, " ");
+            AssertToken(tokens, ref t, TokenTypes.Other, "àë");
+            AssertToken(tokens, ref t, TokenTypes.Whitespace, " ");
+            AssertToken(tokens, ref t, TokenTypes.EOL, "\n");
+            AssertToken(tokens, ref t, TokenTypes.EOF, string.Empty);
+        }
+
+        [TestMethod]
         public void WithStrings_ReturnsStringRuns()
         {
             // Arrange.
-            Setup("A \"simple string\" and 'single quotes' and `back-ticks`.");
+            var lexer = new Lexer("A \"simple string\" and 'single quotes' and `back-ticks`.");
 
             // Act.
             var tokens = lexer.GetTokens();
@@ -72,7 +90,7 @@ namespace Luthor.Tests
         public void WithStrings_FromREADME_ReturnsStringRuns()
         {
             // Arrange.
-            Setup("Sample text.\nAcross 3 lines.\nWith a \"string\".");
+            var lexer = new Lexer("Sample text.\nAcross 3 lines.\nWith a \"multi 'word' string\".");
 
             // Act.
             var tokens = lexer.GetTokens();
@@ -95,7 +113,7 @@ namespace Luthor.Tests
             AssertToken(tokens, ref t, TokenTypes.Whitespace, " ");
             AssertToken(tokens, ref t, TokenTypes.Letters, "a");
             AssertToken(tokens, ref t, TokenTypes.Whitespace, " ");
-            AssertToken(tokens, ref t, TokenTypes.String, "\"string\"");
+            AssertToken(tokens, ref t, TokenTypes.String, "\"multi 'word' string\"");
             AssertToken(tokens, ref t, TokenTypes.Symbols, ".");
             AssertToken(tokens, ref t, TokenTypes.EOF, string.Empty);
         }
@@ -104,7 +122,7 @@ namespace Luthor.Tests
         public void WithUnterminatedString_ReturnsStringRun()
         {
             // Arrange.
-            Setup("An 'unterminated string");
+            var lexer = new Lexer("An 'unterminated string");
 
             // Act.
             var tokens = lexer.GetTokens();
@@ -121,7 +139,7 @@ namespace Luthor.Tests
         public void WithString_WithLineBreak_ReturnsStringRunIncludingLineBreak()
         {
             // Arrange.
-            Setup("A 'string with a \nnew line embedded'.");
+            var lexer = new Lexer("A 'string with a \nnew line embedded'.");
 
             // Act.
             var tokens = lexer.GetTokens();
@@ -139,7 +157,7 @@ namespace Luthor.Tests
         public void WithString_WithEmbeddedStrings_ReturnsSingleStringRun()
         {
             // Arrange.
-            Setup("A \"simple string `with` 'embedded' strings\".");
+            var lexer = new Lexer("A \"simple string `with` 'embedded' strings\".");
 
             // Act.
             var tokens = lexer.GetTokens();
@@ -157,7 +175,7 @@ namespace Luthor.Tests
         public void WithWhitespace_RequestCompressed_ReturnsCompressed()
         {
             // Arrange.
-            Setup("One two   three  \t \r four");
+            var lexer = new Lexer("One two   three  \t \r four");
 
             // Act.
             var tokens = lexer.GetTokens(true);
@@ -171,6 +189,38 @@ namespace Luthor.Tests
             AssertToken(tokens, ref t, TokenTypes.Letters, "three");
             AssertToken(tokens, ref t, TokenTypes.Whitespace, " ");
             AssertToken(tokens, ref t, TokenTypes.Letters, "four");
+            AssertToken(tokens, ref t, TokenTypes.EOF, string.Empty);
+        }
+
+        [TestMethod]
+        public void WithCustomTokenCharacters_ReturnsCorrectTokens()
+        {
+            // Arrange.
+            var lexer = new Lexer("11 22 3--3 44 555 ëë \n")
+            {
+                Chars = "1",
+                Digits = "2",
+                Quotes = "3",      // "--" is written as 3--3
+                Symbols = "4",
+                Whitespace = "5",  // Three spaces are written as 555
+            };
+
+            // Act.
+            var tokens = lexer.GetTokens();
+
+            // Assert.
+            var t = 0;
+            AssertToken(tokens, ref t, TokenTypes.Letters, "11");
+            AssertToken(tokens, ref t, TokenTypes.Other, " ");
+            AssertToken(tokens, ref t, TokenTypes.Digits , "22");
+            AssertToken(tokens, ref t, TokenTypes.Other, " ");
+            AssertToken(tokens, ref t, TokenTypes.String, "3--3");
+            AssertToken(tokens, ref t, TokenTypes.Other, " ");
+            AssertToken(tokens, ref t, TokenTypes.Symbols, "44");
+            AssertToken(tokens, ref t, TokenTypes.Other, " ");
+            AssertToken(tokens, ref t, TokenTypes.Whitespace, "555");
+            AssertToken(tokens, ref t, TokenTypes.Other, " ëë ");
+            AssertToken(tokens, ref t, TokenTypes.EOL, "\n");
             AssertToken(tokens, ref t, TokenTypes.EOF, string.Empty);
         }