+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGE LOG
 
+- Version 2.1.0
+    - Support custom token characters
+    - *Other* tokens now form runs
+        - Before, runs were separate 1-character tokens
 - Version 2.0.0
     - Make the scanner a hidden implementation detail
     - Switch to using MSTest and refactor/tidy tests